Crate canopy

Re-exports

pub use state::StatefulNode;

Modules

backend

commands

cursor

event

geom

inspector

path

script

style

tutils

widgets

Structs

Binder

Binder provides an ergonomic way to specify a set of key bindings using a builder patttern.

Canopy

NodeName

A node name, which consists of lowercase ASCII alphanumeric characters, plus underscores.

NodeState

An opaque structure that Canopy uses to track node state. Each Node has to keep a NodeState structure, and offer it up through the Node::state() method on request.

ParseError

Render

A Render handler for a single node. A new instance is passed to each invocation of Node::render().

Root

A Root node that lives at the base of a Canopy app. It manages modal windows, houses the Inspector and exposes a set of built-in functions.

ViewPort

A ViewPort manages the size of a node and its projection onto the screen.

Enums

Error

Locate

Outcome

Walk

Walk is the return value from traversal closures.

Traits

Core

The API exposed to nodes by Canopy.

DefaultBindings

Loader

A trait that allows widgets to perform recursive initialization of themselves and their children. The most common use for this trait is to load the command sets from a node tree.

Node

Nodes are the basic building-blocks of a Canopy UI. They are composed in a tree, with each node responsible for managing its own children.

StatefulNode

The interface implemented by all nodes that track state.

Functions

fit

Adjust a node so that it fits a viewport. This fits the node to the viewport’s virtual size, then adjusts the node’s view to place as much of it within the viewport’s screen rectangle as possible.

frame

Adjust a node so that viewport’s screen rectangle frames it with a given margin. Fits the child to the viewport screen rect minus the border margin, then adjusts the node’s view to place as much of of it on screen as possible. This function returns a Frame object that can be used to draw a border around the node.

locate

Calls a closure on the root node under (x, y), then recurses up the tree to all children falling under the same point. The function returns the last node that the closure returned a value for, either with Locate::Match (continuing taversal) or Locate::Stop(stopping traversal). Hidden nodes and nodes that do not contain the location point are skipped.

node_at

Find the ID of the leaf node at a given point.

node_path

Return the node path for a specified node id, relative to the specified root.

place

Place a node in a given screen rectangle. This fits the node to the region, and updates its viewport.

postorder

A postorder traversal of the nodes under e.

preorder

Walk::Skip prunes all children of the current node from the traversal.Walk::Handle stops the traversal and the contained value is returned.Any error return stops the traversal and the error is returned.

walk_to_root

Call a closure on the node with the specified id, and all its ancestors to the specified root.

Type Definitions

NodeId

Result

Attribute Macros

command

Mark a method as a command. This macro should be used to decorate methods in an impl block that uses the derive_commands macro. A number of optional arguments can be passed:

derive_commands

Derive an implementation of the CommandNode trait. This macro should be added to the impl block of a struct. All methods that are annotated with command are added as commands, with their doc comments as the command documentation.

Derive Macros

StatefulNode

Derive an implementation of the StatefulNode trait for a struct. The struct should have a self.state attribute of type NodeState.