canopy::backend

Trait BackendControl

pub trait BackendControl: Debug {
    // Required methods
    fn start(&mut self) -> Result<()>;
    fn stop(&mut self) -> Result<()>;

    // Provided method
    fn exit(&mut self, code: i32) -> ! { ... }
}

A handle for controlling our rendering back-end. The primary use is to suspend and resume rendering to permit us to fork out to another process that wants to control the terminal - for example, spawning an external editor.

Required Methods

fn start(&mut self) -> Result<()>

Start the backend renderer.

fn stop(&mut self) -> Result<()>

Stop the backend renderer, releasing control of the terminal.

Provided Methods

fn exit(&mut self, code: i32) -> !

Stop the render backend and exit the process.