canopy

Module geom

Structs

Coverage

Coverage accumulates Rects that have been drawn and calculates which portions of an enclosing area remain un-covered. This allows us to calculate which areas of a Node screen rectangle remain need to be cleared after a render sweep.

Expanse

An Expanse is a rectangle that has a width and height but no location. This is useful when we want to deal with Rects abstractly, or when we want to madate that the location of a Rect is (0, 0).

Frame

A frame extracted from a rectangle

Line

A horizontal line, one character high - essentially a Rect with height 1.

LineSegment

An exctent is a directionless one-dimensional line segment.

Point

Rect

A rectangle

Enums

Direction