canopy

Module layout

Helper functions for Node::layout implementations.

Functions

fit

Adjust a node so that it fits a viewport. This fits the node to the viewport’s screen rectangle, then adjusts the node’s view to place as much of it within the viewport’s screen rectangle as possible.

frame

Adjust a node so that viewport’s screen rectangle frames it with a given margin. Fits the child to the viewport screen rect minus the border margin, then adjusts the node’s view to place as much of of it on screen as possible. This function returns a Frame object that can be used to draw a border around the node.

place

Place a node in a given screen rectangle. This fits the node to the region, and updates its viewport.