canopy::layout

Struct Size

pub struct Size<T> {
    pub width: T,
    pub height: T,
}

Size with width and height.

Fields
width: T

Width component.

height: T

Height component.

Implementations

impl<T> Size<T>

pub fn new(width: T, height: T) -> Self

Create a new size with the given width and height.

impl Size<u32>

pub const ZERO: Self

Zero size.

pub fn main(self, direction: Direction) -> u32

Size along the main axis.

pub fn cross(self, direction: Direction) -> u32

Size along the cross axis.

pub fn from_main_cross(direction: Direction, main: u32, cross: u32) -> Self

Construct a size from main and cross axis values.