canopy::node

Trait Node

pub trait Node: StatefulNode + CommandNode {
    // Provided methods
    fn force_render(&self, c: &dyn Context) -> bool { ... }
    fn cursor(&self) -> Option<Cursor> { ... }
    fn accept_focus(&mut self) -> bool { ... }
    fn children(
        &mut self,
        f: &mut dyn FnMut(&mut dyn Node) -> Result<(), Error>,
    ) -> Result<(), Error> { ... }
    fn handle_key(
        &mut self,
        c: &mut dyn Context,
        k: Key,
    ) -> Result<EventOutcome, Error> { ... }
    fn handle_mouse(
        &mut self,
        c: &mut dyn Context,
        k: MouseEvent,
    ) -> Result<EventOutcome, Error> { ... }
    fn poll(&mut self, c: &mut dyn Context) -> Option<Duration> { ... }
    fn layout(&mut self, l: &Layout, target: Expanse) -> Result<(), Error> { ... }
    fn render(
        &mut self,
        c: &dyn Context,
        r: &mut Render<'_>,
    ) -> Result<(), Error> { ... }
}

Nodes are the basic building-blocks of a Canopy UI. They are composed in a tree, with each node responsible for managing its own children.

Provided Methods

fn force_render(&self, c: &dyn Context) -> bool

Force the node to render in the next sweep. Over-riding this method should only be needed rarely, for instance when a container node (e.g. a frame) needs to redraw if a child node changes.

fn cursor(&self) -> Option<Cursor>

Called for each node on the focus path, after each render sweep. The first node that returns a cursor::Cursor specification controls the cursor. If no node returns a cursor, cursor display is disabled.

fn accept_focus(&mut self) -> bool

Attempt to focus this node. If the node accepts focus, it should return true, and if not return false. The default implementation returns false.

fn children( &mut self, f: &mut dyn FnMut(&mut dyn Node) -> Result<(), Error>, ) -> Result<(), Error>

Call a closure on this node’s children. If any child handler returns an error, processing terminates without visiting the remaining children. The order in which nodes are processed should match intuitive next/previous relationships. The default implementation assumes this node has no children, and just returns.

fn handle_key( &mut self, c: &mut dyn Context, k: Key, ) -> Result<EventOutcome, Error>

Handle a key input event. This event is only called for nodes that are on the focus path. The default implementation ignores input.

fn handle_mouse( &mut self, c: &mut dyn Context, k: MouseEvent, ) -> Result<EventOutcome, Error>

Handle a mouse input event. The default implementation ignores mouse input.

fn poll(&mut self, c: &mut dyn Context) -> Option<Duration>

The scheduled poll endpoint. This function is called for every node the first time it is seen during the pre-render sweep. Each time the function returns a duration, a subsequent call is scheduled. If the function returns None, the poll function is never called again. The default implementation returns None.

fn layout(&mut self, l: &Layout, target: Expanse) -> Result<(), Error>

Re-compute the size and view of the node to display in the target area. In practice, nodes will either constrain themselves based on the width or the height of the target area, or neither, but not both. The resulting size may be smaller or larger than the target. If non-trivial computation is done (e.g. reflowing text), it should be cached for use by future calls.

fn render(&mut self, c: &dyn Context, r: &mut Render<'_>) -> Result<(), Error>

Render this widget, only drawing itself with reference to its own canvas.