canopy

Module prelude

Convenience re-exports for common Canopy types.

Re-exports

pub use crate::Binder;

pub use crate::Canopy;

pub use crate::ChildKey;

pub use crate::Context;

pub use crate::Loader;

pub use crate::NodeId;

pub use crate::ReadContext;

pub use crate::Slot;

pub use crate::TypedId;

pub use crate::Widget;

pub use crate::error;

pub use crate::layout::Align;

pub use crate::layout::Constraint;

pub use crate::layout::Direction;

pub use crate::layout::Display;

pub use crate::layout::Layout;

pub use crate::layout::MeasureConstraints;

pub use crate::layout::Measurement;

pub use crate::layout::Size;

pub use crate::layout::Sizing;

Modules

mouse

Mouse event types.

Macros

key

Define a typed key for keyed children.

Structs

Expanse

An Expanse is a rectangle that has a width and height but no location. This is useful when we want to deal with Rects abstractly, or when we want to madate that the location of a Rect is (0, 0).

Key

A keystroke along with modifiers. A keystroke along with modifiers.

NodeName

A node name, which consists of lowercase ASCII alphanumeric characters, plus underscores.

Point

A 2D point in integer cell coordinates.

Rect

A rectangle

Render

A renderer that only renders to a specific rectangle within the target terminal buffer.

Enums

Event

This enum represents all the event types that drive the application.

Type Aliases

Result

Common result alias for Canopy operations.