Struct canopy::Canopy

pub struct Canopy {
    pub style: StyleMap,
    /* private fields */
}

Fields
style: StyleMap
Implementations

impl Canopy

pub fn new() -> Self

pub fn register_backend<T: BackendControl + 'static>(&mut self, be: T)

pub fn run_script( &mut self, root: &mut dyn Node, node_id: NodeId, sid: ScriptId ) -> Result<()>

pub fn bind_mouse<K>( &mut self, mouse: K, path_filter: &str, script: &str ) -> Result<()>where Mouse: From<K>,

Bind a mouse action in the global mode with a given path filter to a script.

pub fn bind_mode_mouse<K>( &mut self, mouse: K, mode: &str, path_filter: &str, script: &str ) -> Result<()>where Mouse: From<K>,

Bind a mouse action in a specified mode with a given path filter to a script.

pub fn bind_key<K>( &mut self, key: K, path_filter: &str, script: &str ) -> Result<()>where Key: From<K>,

Bind a key in the global mode, with a given path filter to a script.

pub fn bind_mode_key<K>( &mut self, key: K, mode: &str, path_filter: &str, script: &str ) -> Result<()>where Key: From<K>,

Bind a key within a given mode, with a given path filter to a script.

pub fn add_commands<T: CommandNode>(&mut self)

Load the commands from a command node using the default node name derived from the name of the struct.

pub fn print_command_table(&self, w: &mut dyn Write) -> Result<()>

Output a formatted table of commands to a writer.

pub fn node_focus_path_changed(&self, n: &dyn Node) -> bool

Has the focus path status of this node changed since the last render sweep?