canopy

Struct ViewPort

pub struct ViewPort { /* private fields */ }

A ViewPort manages the size of a node and its projection onto the screen. In many ways, this is the core of Canopy, and the viewport structure and its constraints determines many aspects of Canopyโs layout and rendering behavior.

A Canopy app is a tree of nested ViewPorts, with each maintaining a set of constraints with reference to its parent and children.

Implementations

impl ViewPort

pub fn new( canvas: impl Into<Expanse>, view: impl Into<Rect>, position: impl Into<Point>, ) -> Result<ViewPort, Error>

Create a new View with the given outer and inner rectangles. The view rectangle must be fully contained within the outer rectangle.

pub fn position(&self) -> Point

Position of this ViewPortโs view within the parent canvas.

pub fn view(&self) -> Rect

This viewportโs view rectangle, relative to our own canvas.

pub fn canvas(&self) -> Expanse

The canvas size for this viewport.

pub fn set_position(&mut self, p: Point)

Set the viewport position. The caller must provide the parentโs position and canvas rectangle so that we can verify the new position stays within bounds.

pub fn set_canvas(&mut self, sz: Expanse)

Update the canvas size for this viewport, clamping the current view to remain within the new canvas.

pub fn set_view(&mut self, view: Rect)

Set the visible view rectangle, clamped so that it always falls within the current canvas.

pub fn scroll_right(&mut self)

Scroll the view right by one line.

pub fn screen_rect(&self) -> Rect

Absolute rectangle for the screen region the node is being projected onto.

pub fn fit_size(&mut self, size: Expanse, view_size: Expanse)

Set the node size and the target view size at the same time. We try to retain the old view position, but shift and resize it to be within the view if necessary.

pub fn vactive(&self, margin: Rect) -> Result<Option<(Rect, Rect, Rect)>, Error>

Calculates the (pre, active, post) rectangles needed to draw a vertical scroll bar for this viewport in the specified margin rect (usually a right or left vertical margin). Returns None if no scroll bar is needed.

pub fn hactive(&self, margin: Rect) -> Result<Option<(Rect, Rect, Rect)>, Error>

Calculates the (pre, active, post) rectangles needed to draw a horizontal scroll bar for this viewport in the specified margin rect (usually a bottom horizontal margin). Returns None if no scroll bar is needed.