canopy::style::effects

Trait StyleEffect

pub trait StyleEffect:
    Send
    + Sync
    + Debug {
    // Required methods
    fn apply(&self, style: Style) -> Style;
    fn box_clone(&self) -> Box<dyn StyleEffect>;
}

A style transformation that can be applied during rendering.

Effects are stacked and applied in order during render traversal. They inherit through the tree unless explicitly cleared.

Required Methods

fn apply(&self, style: Style) -> Style

Apply this effect to a style, returning the transformed style.

fn box_clone(&self) -> Box<dyn StyleEffect>

Clone this effect into a boxed trait object.