canopy::style

Struct PartialStyle

pub struct PartialStyle {
    pub fg: Option<Color>,
    pub bg: Option<Color>,
    pub attrs: Option<AttrSet>,
}

A possibly partial style specification, which is stored in a StyleManager. Partial styles are completely resolved during the style resolution process.

Fields
fg: Option<Color>

Optional foreground color.

bg: Option<Color>

Optional background color.

attrs: Option<AttrSet>

Optional attributes.

Implementations

impl PartialStyle

pub fn fg(fg: Color) -> Self

Create a new PartialStyle with only a foreground color.

pub fn bg(bg: Color) -> Self

Create a new PartialStyle with only a background color.

pub fn attrs(attrs: AttrSet) -> Self

Create a new PartialStyle with only attributes.

pub fn resolve(&self) -> Style

Resolve the partial style into a full style.

pub fn with_fg(self, fg: Color) -> Self

Set the foreground color.

pub fn with_bg(self, bg: Color) -> Self

Set the background color.

pub fn with_attr(self, attr: Attr) -> Self

Add a single attribute.

pub fn with_attrs(self, attrs: AttrSet) -> Self

Replace the attributes set.

pub fn join(&self, other: &Self) -> Self

Merge two partial styles.

pub fn is_complete(&self) -> bool

Return true if all components are set.