canopy

Trait CommandNode

pub trait CommandNode: StatefulNode {
    // Required methods
    fn commands() -> Vec<CommandSpec>
       where Self: Sized;
    fn dispatch(
        &mut self,
        c: &mut dyn Context,
        cmd: &CommandInvocation,
    ) -> Result<ReturnValue, Error>;
}

The CommandNode trait is implemented by all Nodes to expose the set of supported commands. With rare exceptions, this is done with the commands macro.

Required Methods

fn commands() -> Vec<CommandSpec>
where Self: Sized,

Return a list of commands for this node.

fn dispatch( &mut self, c: &mut dyn Context, cmd: &CommandInvocation, ) -> Result<ReturnValue, Error>

Dispatch a command to this node.