Trait canopy::Core

pub trait Core {

    // Required methods
    fn is_on_focus_path(&self, n: &mut dyn Node) -> bool;
    fn is_focused(&self, n: &dyn Node) -> bool;
    fn focus_area(&self, root: &mut dyn Node) -> Option<Rect>;
    fn focus_depth(&self, n: &mut dyn Node) -> usize;
    fn focus_down(&mut self, root: &mut dyn Node) -> Result<Outcome>;
    fn focus_first(&mut self, root: &mut dyn Node) -> Result<Outcome>;
    fn focus_left(&mut self, root: &mut dyn Node) -> Result<Outcome>;
    fn focus_next(&mut self, root: &mut dyn Node) -> Result<Outcome>;
    fn focus_path(&self, root: &mut dyn Node) -> Path;
    fn focus_prev(&mut self, root: &mut dyn Node) -> Result<Outcome>;
    fn focus_right(&mut self, root: &mut dyn Node) -> Result<Outcome>;
    fn focus_up(&mut self, root: &mut dyn Node) -> Result<Outcome>;
    fn needs_render(&self, n: &dyn Node) -> bool;
    fn set_focus(&mut self, n: &mut dyn Node);
    fn focus_dir(
        &mut self,
        root: &mut dyn Node,
        dir: Direction
    ) -> Result<Outcome>;
    fn taint(&mut self, n: &mut dyn Node);
    fn taint_tree(&mut self, e: &mut dyn Node);
    fn start(&mut self) -> Result<()>;
    fn stop(&mut self) -> Result<()>;
    fn exit(&mut self, code: i32) -> !;

}

The API exposed to nodes by Canopy.

Required Methods

fn is_on_focus_path(&self, n: &mut dyn Node) -> bool

Is the specified node on the focus path? A node is on the focus path if it has focus, or if it’s the ancestor of a node with focus.

fn is_focused(&self, n: &dyn Node) -> bool

Does the node have focus?

fn focus_area(&self, root: &mut dyn Node) -> Option<Rect>

Get the Rect of the screen area that currently has focus.

fn focus_depth(&self, n: &mut dyn Node) -> usize

What is the depth of the specified node in the focus path?

fn focus_down(&mut self, root: &mut dyn Node) -> Result<Outcome>

Move focus downward of the currently focused node within the subtree at root.

fn focus_first(&mut self, root: &mut dyn Node) -> Result<Outcome>

Focus the first node that accepts focus in the pre-order traversal of the subtree at root.

fn focus_left(&mut self, root: &mut dyn Node) -> Result<Outcome>

Move focus to the left of the currently focused node within the subtree at root.

fn focus_next(&mut self, root: &mut dyn Node) -> Result<Outcome>

Focus the next node in the pre-order traversal of root. If no node with focus is found, we focus the first node we can find instead.

fn focus_path(&self, root: &mut dyn Node) -> Path

Return the focus path for the subtree under root.

fn focus_prev(&mut self, root: &mut dyn Node) -> Result<Outcome>

Focus the previous node in the pre-order traversal of root. If no node with focus is found, we focus the first node we can find instead.

fn focus_right(&mut self, root: &mut dyn Node) -> Result<Outcome>

Move focus to right of the currently focused node within the subtree at root.

fn focus_up(&mut self, root: &mut dyn Node) -> Result<Outcome>

Move focus upward of the currently focused node within the subtree at root.

fn needs_render(&self, n: &dyn Node) -> bool

Does the node need to render in the next sweep? This checks if the node is currently hidden, and if not, signals that we should render if the node is tainted, its focus status has changed, or if it is forcing a render.

fn set_focus(&mut self, n: &mut dyn Node)

Focus a node.

fn focus_dir(&mut self, root: &mut dyn Node, dir: Direction) -> Result<Outcome>

Move focus in a specified direction within the subtree at root.

fn taint(&mut self, n: &mut dyn Node)

Taint a node to signal that it should be re-rendered.

fn taint_tree(&mut self, e: &mut dyn Node)

Taint the entire subtree under a node.

fn start(&mut self) -> Result<()>

Start the backend renderer.

fn stop(&mut self) -> Result<()>

Stop the backend renderer, releasing control of the terminal.

fn exit(&mut self, code: i32) -> !

Stop the render backend and exit the process.