canopy

Module tree

Tree traversal utilities. Utilities for working with a Canopy node tree.

Enums

Locate

Result of a locate traversal.

Walk

Walk is the return value from traversal closures.

Functions

locate

Calls a closure on the root node under (x, y), then recurses up the tree to all children falling under the same point. The function returns the last node that the closure returned a value for, either with Locate::Match (continuing taversal) or Locate::Stop(stopping traversal). Hidden nodes and nodes that do not contain the location point are skipped.

node_at

Find the ID of the leaf node at a given point.

node_path

Return the node path for a specified node id, relative to the specified root.

postorder

A postorder traversal of the nodes under e.

preorder

Walk::Skip prunes all children of the current node from the traversal.Walk::Handle stops the traversal and the contained value is returned.Any error return stops the traversal and the error is returned.

walk_focus_path_e

Call a closure on the currently focused node and all its ancestors to the root. If the closure returns Walk::Handle, traversal stops. Handle::Skip is ignored.

walk_to_root

Call a closure on the node with the specified id, and all its ancestors to the specified root.