canopy::widgets::editor::core

Struct Core

pub struct Core { /* private fields */ }

The editor Core exposes the operations that can be performed on a text buffer. Itโs a facade over the state, with added operations to support a redo/undo stack.

Implementations

impl Core

pub fn new(text: &str) -> Self

Construct a new editor core from the provided text.

pub fn undo(&mut self) -> bool

Undo an operation. Return true if an operation was undone, false if the history is empty.

pub fn redo(&mut self) -> bool

Redo an operation. Returne true if an operation was redone, false if redo history is empty.

pub fn insert_text(&mut self, text: &str)

Insert text at the current cursor position.

pub fn delete<T>(&mut self, start: T, end: T)
where T: Into<InsertPos>,

Delete text in a given range.

pub fn window_text(&self) -> Vec<Option<&str>>

Return the visible wrapped lines for the current window.

pub fn wrapped_height(&self) -> usize

Return the total wrapped line count for the document.

pub fn resize_window(&mut self, width: usize, height: usize)

Resize the visible window to a new width and height.

pub fn cursor_position(&self) -> Option<Point>

Return the cursor position in screen coordinates.

pub fn cursor_shift(&mut self, n: isize)

Move the cursor within the current chunk, moving to the next or previous wrapped line if needed. Wonโt move to the next chunk.

pub fn cursor_shift_chunk(&mut self, n: isize)

Move the up or down in the chunk list.

pub fn cursor_shift_lines(&mut self, n: isize)

Move the up or down along wrapped lines.