Crate canopy

A library for building terminal UIs.

Re-exports

pub use geom::Expanse;

pub use layout::*;

Modules

backend

commands

cursor

event

geom

inspector

layout

Helper functions for Node::layout implementations.

path

script

style

tree

Utilities for working with a Canopy node tree.

widgets

Structs

Binder

Binder provides an ergonomic way to specify a set of key bindings using a builder patttern.

Canopy

NodeId

A unique ID for a node.

NodeName

A node name, which consists of lowercase ASCII alphanumeric characters, plus underscores.

NodeState

An opaque structure that Canopy uses to track node state. Each Node has to keep a NodeState structure, and offer it up through the Node::state() method on request.

Render

The interface used to render to the screen. It is only accessible in Node::render.

Root

A Root node that lives at the base of a Canopy app. It manages modal windows, houses the Inspector and exposes a set of built-in functions.

ViewPort

A ViewPort manages the size of a node and its projection onto the screen.

Enums

Error

EventOutcome

Was an event handled or ignored?

Traits

Context

The API exposed to nodes by Canopy.

DefaultBindings

Loader

A trait that allows widgets to perform recursive initialization of themselves and their children. The most common use for this trait is to load the command sets from a node tree.

Node

Nodes are the basic building-blocks of a Canopy UI. They are composed in a tree, with each node responsible for managing its own children.

StatefulNode

The interface implemented by all nodes that track state.

Type Aliases

Result

Attribute Macros

command

Mark a method as a command. This macro should be used to decorate methods in an impl block that uses the derive_commands macro. A number of optional arguments can be passed:

derive_commands

Derive an implementation of the CommandNode trait. This macro should be added to the impl block of a struct. All methods that are annotated with command are added as commands, with their doc comments as the command documentation.

Derive Macros

StatefulNode

Derive an implementation of the StatefulNode trait for a struct. The struct should have a self.state attribute of type NodeState.