canopy

Module commands

Command definition and dispatch.

Structs

CommandInvocation

A parsed command invocation.

CommandSet

Collection of available commands keyed by name.

CommandSpec

CommandDefinition encapsulates the definition of a command that can be performed on a Node. Commands are used for key bindings, mouse actions and general automation.

ReturnSpec

The return type of a command.

Enums

ArgTypes

Supported argument types for command signatures.

Args

Runtime command argument values.

ReturnTypes

The return type of a command.

ReturnValue

Runtime return values from command dispatch.

Traits

CommandNode

The CommandNode trait is implemented by all Nodes to expose the set of supported commands. With rare exceptions, this is done with the commands macro.

Functions

dispatch

Dispatch a command relative to a node. This searches the node tree for a matching node::command in the following order: - A pre-order traversal of the current node subtree - The path from the current node to the root