canopy::state

Trait StatefulNode

pub trait StatefulNode {
    // Required methods
    fn name(&self) -> NodeName;
    fn state(&self) -> &NodeState;
    fn state_mut(&mut self) -> &mut NodeState;

    // Provided methods
    fn hide(&mut self) { ... }
    fn unhide(&mut self) { ... }
    fn is_hidden(&self) -> bool { ... }
    fn vp(&self) -> ViewPort { ... }
    fn id(&self) -> NodeId { ... }
    fn is_initialized(&self) -> bool { ... }
}

The interface implemented by all nodes that track state.

Required Methods

fn name(&self) -> NodeName

The name of this node, used for debugging and command dispatch.

fn state(&self) -> &NodeState

Get a reference to the node’s state object.

fn state_mut(&mut self) -> &mut NodeState

Get a mutable reference to the node’s state object.

Provided Methods

fn hide(&mut self)

Hides the element and all its descendants from rendering. The nodes are still included in the tree.

fn unhide(&mut self)

Hides the element

fn is_hidden(&self) -> bool

Is this element hidden?

fn vp(&self) -> ViewPort

Get the node’s ViewPort.

fn id(&self) -> NodeId

A unique ID for this node.

fn is_initialized(&self) -> bool

Has this node been initialized? That is, has its poll function been called for the first time to schedule future polls.