canopy

Trait Context

pub trait Context {

    // Required methods
    fn needs_render(&self, n: &dyn Node) -> bool;
    fn is_on_focus_path(&self, n: &mut dyn Node) -> bool;
    fn is_focused(&self, n: &dyn Node) -> bool;
    fn focus_area(&self, root: &mut dyn Node) -> Option<Rect>;
    fn focus_depth(&self, n: &mut dyn Node) -> usize;
    fn focus_down(&mut self, root: &mut dyn Node);
    fn focus_first(&mut self, root: &mut dyn Node);
    fn focus_left(&mut self, root: &mut dyn Node);
    fn focus_next(&mut self, root: &mut dyn Node);
    fn focus_path(&self, root: &mut dyn Node) -> Path;
    fn focus_prev(&mut self, root: &mut dyn Node);
    fn focus_right(&mut self, root: &mut dyn Node);
    fn focus_up(&mut self, root: &mut dyn Node);
    fn set_focus(&mut self, n: &mut dyn Node);
    fn focus_dir(&mut self, root: &mut dyn Node, dir: Direction);
    fn taint(&mut self, n: &mut dyn Node);
    fn taint_tree(&mut self, e: &mut dyn Node);
    fn start(&mut self) -> Result<()>;
    fn stop(&mut self) -> Result<()>;
    fn exit(&mut self, code: i32) -> !;

    // Provided methods
    fn scroll_to(&self, n: &mut dyn Node, x: u16, y: u16) { ... }
    fn scroll_by(&self, n: &mut dyn Node, x: i16, y: i16) { ... }
    fn page_up(&self, n: &mut dyn Node) { ... }
    fn page_down(&self, n: &mut dyn Node) { ... }
    fn scroll_up(&self, n: &mut dyn Node) { ... }
    fn scroll_down(&self, n: &mut dyn Node) { ... }
    fn scroll_left(&self, n: &mut dyn Node) { ... }
    fn scroll_right(&self, n: &mut dyn Node) { ... }

}

The API exposed to nodes by Canopy.

Required Methods

fn needs_render(&self, n: &dyn Node) -> bool

Does the node need to render in the next sweep? This checks if the node is currently hidden, and if not, signals that we should render if the node is tainted, its focus status has changed, or if it is forcing a render.

fn is_on_focus_path(&self, n: &mut dyn Node) -> bool

Is the specified node on the focus path? A node is on the focus path if it has focus, or if it’s the ancestor of a node with focus.

fn is_focused(&self, n: &dyn Node) -> bool

Does the node have focus?

fn focus_area(&self, root: &mut dyn Node) -> Option<Rect>

Get the Rect of the screen area that currently has focus.

fn focus_depth(&self, n: &mut dyn Node) -> usize

What is the depth of the specified node in the focus path?

fn focus_down(&mut self, root: &mut dyn Node)

Move focus downward of the currently focused node within the subtree at root.

fn focus_first(&mut self, root: &mut dyn Node)

Focus the first node that accepts focus in the pre-order traversal of the subtree at root.

fn focus_left(&mut self, root: &mut dyn Node)

Move focus to the left of the currently focused node within the subtree at root.

fn focus_next(&mut self, root: &mut dyn Node)

Focus the next node in the pre-order traversal of root. If no node with focus is found, we focus the first node we can find instead.

fn focus_path(&self, root: &mut dyn Node) -> Path

Return the focus path for the subtree under root.

fn focus_prev(&mut self, root: &mut dyn Node)

Focus the previous node in the pre-order traversal of root. If no node with focus is found, we focus the first node we can find instead.

fn focus_right(&mut self, root: &mut dyn Node)

Move focus to right of the currently focused node within the subtree at root.

fn focus_up(&mut self, root: &mut dyn Node)

Move focus upward of the currently focused node within the subtree at root.

fn set_focus(&mut self, n: &mut dyn Node)

Focus a node.

fn focus_dir(&mut self, root: &mut dyn Node, dir: Direction)

Move focus in a specified direction within the subtree at root.

fn taint(&mut self, n: &mut dyn Node)

Taint a node to signal that it should be re-rendered.

fn taint_tree(&mut self, e: &mut dyn Node)

Taint the entire subtree under a node.

fn start(&mut self) -> Result<()>

Start the backend renderer.

fn stop(&mut self) -> Result<()>

Stop the backend renderer, releasing control of the terminal.

fn exit(&mut self, code: i32) -> !

Stop the render backend and exit the process.

Provided Methods

fn scroll_to(&self, n: &mut dyn Node, x: u16, y: u16)

Scroll the view to the specified position. The view is clamped within the outer rectangle.

fn scroll_by(&self, n: &mut dyn Node, x: i16, y: i16)

Scroll the view by the given offsets. The view rectangle is clamped within the outer rectangle.

fn page_up(&self, n: &mut dyn Node)

Scroll the view up by the height of the view rectangle.

fn page_down(&self, n: &mut dyn Node)

Scroll the view down by the height of the view rectangle.

fn scroll_up(&self, n: &mut dyn Node)

Scroll the view up by one line.

fn scroll_down(&self, n: &mut dyn Node)

Scroll the view down by one line.

fn scroll_left(&self, n: &mut dyn Node)

Scroll the view left by one line.

fn scroll_right(&self, n: &mut dyn Node)

Scroll the view right by one line.