canopy

Trait Context

pub trait Context: ViewContext {

    // Required methods
    fn set_focus(&mut self, node: NodeId) -> bool;
    fn focus_dir_in(&mut self, root: NodeId, dir: Direction);
    fn focus_first_in(&mut self, root: NodeId);
    fn focus_next_in(&mut self, root: NodeId);
    fn focus_prev_in(&mut self, root: NodeId);
    fn scroll_to(&mut self, x: u32, y: u32) -> bool;
    fn scroll_by(&mut self, x: i32, y: i32) -> bool;
    fn page_up(&mut self) -> bool;
    fn page_down(&mut self) -> bool;
    fn scroll_up(&mut self) -> bool;
    fn scroll_down(&mut self) -> bool;
    fn scroll_left(&mut self) -> bool;
    fn scroll_right(&mut self) -> bool;
    fn with_layout_of(
        &mut self,
        node: NodeId,
        f: &mut dyn FnMut(&mut Layout),
    ) -> Result<()>;
    fn add(&mut self, widget: Box<dyn Widget>) -> NodeId;
    fn with_widget_mut(
        &mut self,
        node: NodeId,
        f: &mut dyn FnMut(&mut dyn Widget, &mut dyn Context) -> Result<()>,
    ) -> Result<()>;
    fn dispatch_command(
        &mut self,
        cmd: &CommandInvocation,
    ) -> Result<Option<ReturnValue>>;
    fn mount_child_to(&mut self, parent: NodeId, child: NodeId) -> Result<()>;
    fn detach_child_from(&mut self, parent: NodeId, child: NodeId) -> Result<()>;
    fn set_children_of(
        &mut self,
        parent: NodeId,
        children: Vec<NodeId>,
    ) -> Result<()>;
    fn set_hidden_of(&mut self, node: NodeId, hidden: bool) -> bool;
    fn start(&mut self) -> Result<()>;
    fn stop(&mut self) -> Result<()>;
    fn exit(&mut self, code: i32) -> !;
    fn push_effect(
        &mut self,
        node: NodeId,
        effect: Box<dyn StyleEffect>,
    ) -> Result<()>;
    fn clear_effects(&mut self, node: NodeId) -> Result<()>;
    fn set_clear_inherited_effects(
        &mut self,
        node: NodeId,
        clear: bool,
    ) -> Result<()>;
    fn set_style(&mut self, style: StyleMap);

    // Provided methods
    fn focus_dir(&mut self, dir: Direction) { ... }
    fn focus_dir_global(&mut self, dir: Direction) { ... }
    fn focus_first(&mut self) { ... }
    fn focus_first_global(&mut self) { ... }
    fn focus_next(&mut self) { ... }
    fn focus_next_global(&mut self) { ... }
    fn focus_prev(&mut self) { ... }
    fn focus_prev_global(&mut self) { ... }
    fn focus_right(&mut self) { ... }
    fn focus_right_in(&mut self, root: NodeId) { ... }
    fn focus_right_global(&mut self) { ... }
    fn focus_left(&mut self) { ... }
    fn focus_left_in(&mut self, root: NodeId) { ... }
    fn focus_left_global(&mut self) { ... }
    fn focus_up(&mut self) { ... }
    fn focus_up_in(&mut self, root: NodeId) { ... }
    fn focus_up_global(&mut self) { ... }
    fn focus_down(&mut self) { ... }
    fn focus_down_in(&mut self, root: NodeId) { ... }
    fn focus_down_global(&mut self) { ... }
    fn with_layout(&mut self, f: &mut dyn FnMut(&mut Layout)) -> Result<()> { ... }
    fn mount_child(&mut self, child: NodeId) -> Result<()> { ... }
    fn detach_child(&mut self, child: NodeId) -> Result<()> { ... }
    fn set_children(&mut self, children: Vec<NodeId>) -> Result<()> { ... }
    fn set_hidden(&mut self, hidden: bool) -> bool { ... }
    fn hide(&mut self) -> bool { ... }
    fn hide_node(&mut self, node: NodeId) -> bool { ... }
    fn show(&mut self) -> bool { ... }
    fn show_node(&mut self, node: NodeId) -> bool { ... }
    fn current_focus_gen(&self) -> u64 { ... }

}

Mutable context available to widgets during event handling.

Required Methods

fn set_focus(&mut self, node: NodeId) -> bool

Focus a node. Returns true if focus changed.

fn focus_dir_in(&mut self, root: NodeId, dir: Direction)

Move focus in a specified direction within the specified subtree.

fn focus_first_in(&mut self, root: NodeId)

Focus the first node that accepts focus in the specified subtree.

fn focus_next_in(&mut self, root: NodeId)

Focus the next node in the specified subtree.

fn focus_prev_in(&mut self, root: NodeId)

Focus the previous node in the specified subtree.

fn scroll_to(&mut self, x: u32, y: u32) -> bool

Scroll the view to the specified position. Returns true if movement occurred.

fn scroll_by(&mut self, x: i32, y: i32) -> bool

Scroll the view by the given offsets. Returns true if movement occurred.

fn page_up(&mut self) -> bool

Scroll the view up by one page. Returns true if movement occurred.

fn page_down(&mut self) -> bool

Scroll the view down by one page. Returns true if movement occurred.

fn scroll_up(&mut self) -> bool

Scroll the view up by one line. Returns true if movement occurred.

fn scroll_down(&mut self) -> bool

Scroll the view down by one line. Returns true if movement occurred.

fn scroll_left(&mut self) -> bool

Scroll the view left by one line. Returns true if movement occurred.

fn scroll_right(&mut self) -> bool

Scroll the view right by one line. Returns true if movement occurred.

fn with_layout_of( &mut self, node: NodeId, f: &mut dyn FnMut(&mut Layout), ) -> Result<()>

Update the layout for a specific node.

fn add(&mut self, widget: Box<dyn Widget>) -> NodeId

Add a new widget node to the core.

fn with_widget_mut( &mut self, node: NodeId, f: &mut dyn FnMut(&mut dyn Widget, &mut dyn Context) -> Result<()>, ) -> Result<()>

Execute a closure with mutable access to a widget and its node-bound context.

fn dispatch_command( &mut self, cmd: &CommandInvocation, ) -> Result<Option<ReturnValue>>

Dispatch a command relative to this node.

fn mount_child_to(&mut self, parent: NodeId, child: NodeId) -> Result<()>

Attach a child to a specific parent node.

fn detach_child_from(&mut self, parent: NodeId, child: NodeId) -> Result<()>

Detach a child from a specific parent node.

fn set_children_of( &mut self, parent: NodeId, children: Vec<NodeId>, ) -> Result<()>

Replace the children list for a specific parent node.

fn set_hidden_of(&mut self, node: NodeId, hidden: bool) -> bool

Set a specific node’s visibility. Returns true if visibility changed.

fn start(&mut self) -> Result<()>

Start the backend renderer.

fn stop(&mut self) -> Result<()>

Stop the backend renderer, releasing control of the terminal.

fn exit(&mut self, code: i32) -> !

Stop the render backend and exit the process.

fn push_effect( &mut self, node: NodeId, effect: Box<dyn StyleEffect>, ) -> Result<()>

Add an effect to a node that will be applied during rendering. Effects stack and inherit through the tree.

fn clear_effects(&mut self, node: NodeId) -> Result<()>

Clear all effects on a node.

fn set_clear_inherited_effects( &mut self, node: NodeId, clear: bool, ) -> Result<()>

Set whether a node should clear inherited effects before applying local ones.

fn set_style(&mut self, style: StyleMap)

Set the style map to be used for rendering. The style change will be applied before the next render.

Provided Methods

fn focus_dir(&mut self, dir: Direction)

Move focus in a specified direction within the current node’s subtree.

fn focus_dir_global(&mut self, dir: Direction)

Move focus in a specified direction within the entire tree (from root).

fn focus_first(&mut self)

Focus the first node that accepts focus in the current node’s subtree.

fn focus_first_global(&mut self)

Focus the first node that accepts focus in the entire tree (from root).

fn focus_next(&mut self)

Focus the next node in the current node’s subtree.

fn focus_next_global(&mut self)

Focus the next node in the entire tree (from root).

fn focus_prev(&mut self)

Focus the previous node in the current node’s subtree.

fn focus_prev_global(&mut self)

Focus the previous node in the entire tree (from root).

fn focus_right(&mut self)

Move focus to the right within the current node’s subtree.

fn focus_right_in(&mut self, root: NodeId)

Move focus to the right within the specified subtree.

fn focus_right_global(&mut self)

Move focus to the right within the entire tree (from root).

fn focus_left(&mut self)

Move focus to the left within the current node’s subtree.

fn focus_left_in(&mut self, root: NodeId)

Move focus to the left within the specified subtree.

fn focus_left_global(&mut self)

Move focus to the left within the entire tree (from root).

fn focus_up(&mut self)

Move focus upward within the current node’s subtree.

fn focus_up_in(&mut self, root: NodeId)

Move focus upward within the specified subtree.

fn focus_up_global(&mut self)

Move focus upward within the entire tree (from root).

fn focus_down(&mut self)

Move focus downward within the current node’s subtree.

fn focus_down_in(&mut self, root: NodeId)

Move focus downward within the specified subtree.

fn focus_down_global(&mut self)

Move focus downward within the entire tree (from root).

fn with_layout(&mut self, f: &mut dyn FnMut(&mut Layout)) -> Result<()>

Update the layout for the current node.

fn mount_child(&mut self, child: NodeId) -> Result<()>

Attach a child to the current node.

fn detach_child(&mut self, child: NodeId) -> Result<()>

Detach a child from the current node.

fn set_children(&mut self, children: Vec<NodeId>) -> Result<()>

Replace the children list for the current node.

fn set_hidden(&mut self, hidden: bool) -> bool

Set the current node’s visibility. Returns true if visibility changed.

fn hide(&mut self) -> bool

Hide the current node. Returns true if visibility changed.

fn hide_node(&mut self, node: NodeId) -> bool

Hide a specific node. Returns true if visibility changed.

fn show(&mut self) -> bool

Show the current node. Returns true if visibility changed.

fn show_node(&mut self, node: NodeId) -> bool

Show a specific node. Returns true if visibility changed.

fn current_focus_gen(&self) -> u64

Current focus generation counter.

Implementations

impl dyn Context + '_

pub fn with_widget<W, R>( &mut self, node: NodeId, f: impl FnMut(&mut W, &mut dyn Context) -> Result<R>, ) -> Result<R>
where W: Widget + 'static,

Execute a closure with mutable access to a widget of type W.

pub fn with_typed<W, R>( &mut self, node: TypedId<W>, f: impl FnMut(&mut W, &mut dyn Context) -> Result<R>, ) -> Result<R>
where W: Widget + 'static,

Execute a closure with mutable access to a widget using a typed node ID.

pub fn try_with_widget<W, R>( &mut self, node: NodeId, f: impl FnMut(&mut W, &mut dyn Context) -> Result<R>, ) -> Result<Option<R>>
where W: Widget + 'static,

Execute a closure with mutable access to a widget of type W if it matches.

pub fn try_with_typed<W, R>( &mut self, node: TypedId<W>, f: impl FnMut(&mut W, &mut dyn Context) -> Result<R>, ) -> Result<Option<R>>
where W: Widget + 'static,

Execute a closure with mutable access to a widget using a typed node ID if it matches.

pub fn add_orphan<W: Widget + 'static>(&mut self, widget: W) -> NodeId

Add a widget to the core but do not attach it to any parent (orphan).

pub fn add_orphan_typed<W: Widget + 'static>(&mut self, widget: W) -> TypedId<W>

Add a widget to the core and return a typed node identifier (orphan).

pub fn add_child<W: Widget + 'static>(&mut self, widget: W) -> Result<NodeId>

Add a widget as a child of the current node and return the new node ID.

pub fn add_child_to<W: Widget + 'static>( &mut self, parent: NodeId, widget: W, ) -> Result<NodeId>

Add a widget as a child of a specific parent and return the new node ID.

pub fn add_children<I>(&mut self, widgets: I) -> Result<Vec<NodeId>>
where I: IntoIterator<Item = Box<dyn Widget>>,

Add multiple boxed widgets as children of the current node and return their node IDs.

pub fn add_children_to<I>( &mut self, parent: NodeId, widgets: I, ) -> Result<Vec<NodeId>>
where I: IntoIterator<Item = Box<dyn Widget>>,

Add multiple boxed widgets as children of a specific parent and return their node IDs.

pub fn only_child(&self) -> Option<NodeId>

Return the only child of this node, or None if there are no children.

Panics

Panics if there is more than one child.

pub fn suggest_focus_after_remove( &mut self, root: NodeId, removed: NodeId, ) -> Option<NodeId>

Suggest a focus target after removing removed from the subtree rooted at root.